requirements.txt
# === Runtime dependencies (for development & execution) ===
tkinterweb==4.7.0       # embedded HTML/CSS preview
fpdf2==2.7.9             # PDF generator
watchdog==5.0.2          # auto-reload watcher
pywin32==306             # Windows integration (optional but safe)

# === Build / packaging dependencies ===
pyinstaller>=6.0         # build the EXE
weasyprint>=61.0         # optional rich PDF engine
schedule>=1.2.0          # optional background scheduler

tkcalendar==1.6.1
resend==0.7.0
python-dotenv==1.0.0